+++ src/Command/Bake/DocumentCommand.php
@@ -273,6 +273,11 @@ class DocumentCommand extends BakeCommand
     protected function buildOptionParser(ConsoleOptionParser $parser): ConsoleOptionParser
     {
         $parser = $this->_setCommonOptions($parser);
+        $parser->addOption('connection', [
+            'default' => 'mongo',
+            'help' => 'The datasource connection to get data from.',
+        ]);
+
         $parser->setDescription(static::getDescription())
             ->addArgument('name', [
                 'help' => 'Name of the document class to bake (e.g., Article).',